Handle health sensor registry reuse
